requirements.txt
torch>=2.2.0
transformers>=4.40.0
accelerate>=0.29.0
bitsandbytes>=0.43.0
datasets>=2.19.0
peft>=0.10.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pinecone-client>=3.1.0
clip-anytorch>=2.5.0
Pillow>=10.4.0
scipy
tqdm